Delta electron range models implementation refactored. Type "long" replaced by "int" where necessary.

+++ src/AT_RDD.c
@@ -461,8 +461,8 @@ void AT_RDD_f1_parameters(  /* radiation field parameters */
   // PARAMETER 2: Get the maximum electron range (same for all RDD models)
   AT_max_electron_range_m(  n_tmp,
                 E_MeV_u,
-                *material_no,
-                *er_model,
+                (int)(*material_no),
+                (int)(*er_model),
                 &max_electron_range_m);
 
   ////////////////////////////////////////////////////////////////////////////////
